+++ src/page.ts
@@ -5,20 +5,27 @@ import {
     EventTransportMessage,
     EventTransportMessagePayload
 } from 'rempl';
+import { BgToPageMessage, PageToPluginMessage, PublisherInfo } from './types';
 import { createIndicator, genUID } from './helpers';
-import { BgToPageMessage, PageToPluginMessage } from './types';
+
+type PublisherHub = {
+    id: number;
+    connected: boolean;
+    channelId: EventTransportChannelId;
+    publishers: PublisherInfo[];
+};
 
 const DEBUG = false;
 const sessionId = genUID();
 let pluginConnected = false;
 let remplConnected = false;
-let publishers: string[] = [];
-let subscribers: string[] = [];
+let subscribers: PublisherInfo[] = [];
 const debugIndicator = DEBUG ? createIndicator() : null;
 const name = 'rempl-browser-extension-host';
 const connectToName = 'rempl-browser-extension-publisher';
 const inputChannelId: EventTransportChannelId = `${name}/${genUID()}`;
-let outputChannelId: EventTransportChannelId | null = null;
+const publisherHubs = new Map<EventTransportChannelId, PublisherHub>();
+const publishers = new Map<string, PublisherInfo>();
 
 const plugin = chrome.runtime.connect({
     name: 'rempl:page'
@@ -54,9 +61,33 @@ function emitPageEvent(to: EventTransportConnectTo | EventTransportChannelId, pa
     postMessage(message, '*');
 }
 
-function sendToPage(data: EventTransportMessagePayload) {
-    if (outputChannelId) {
-        emitPageEvent(outputChannelId, data);
+function sendToPage(channelId: EventTransportChannelId, data: EventTransportMessagePayload) {
+    emitPageEvent(channelId, data);
+}
+
+function syncPublisherConnections() {
+    const connectedPublishers = new Set(subscribers.map(({ id }) => id));
+
+    for (const hub of publisherHubs.values()) {
+        const { connected, channelId, publishers } = hub;
+        const endpoints = publishers
+            .map(({ id }) => id)
+            .filter((id) => connectedPublishers.has(id));
+        const shouldBeConnected = endpoints.length > 0;
+
+        if (connected !== shouldBeConnected) {
+            hub.connected = shouldBeConnected;
+            if (shouldBeConnected) {
+                sendToPage(channelId, {
+                    type: 'connect',
+                    endpoints
+                });
+            } else {
+                sendToPage(channelId, {
+                    type: 'disconnect'
+                });
+            }
+        }
     }
 }
 
@@ -65,7 +96,7 @@ function handshake(inited: boolean) {
         type: 'handshake',
         initiator: name,
         inited,
-        endpoints: subscribers
+        endpoints: []
     });
 }
 
@@ -81,13 +112,11 @@ plugin.onMessage.addListener((packet: BgToPageMessage) => {
     switch (packet.type) {
         case 'connect':
             if (!pluginConnected && remplConnected) {
+                cancelPluginPublishersSync();
+                syncPublisherConnections();
                 sendToPlugin({
                     type: 'page:connect',
-                    data: [sessionId, publishers]
-                });
-                sendToPage({
-                    type: 'connect',
-                    endpoints: subscribers
+                    data: [sessionId, [...publishers.values()]]
                 });
             }
 
@@ -97,26 +126,32 @@ plugin.onMessage.addListener((packet: BgToPageMessage) => {
             break;
 
         case 'disconnect':
-            if (pluginConnected && remplConnected) {
-                sendToPage({
-                    type: 'disconnect'
-                });
-            }
-
+            subscribers = [];
             pluginConnected = false;
+            syncPublisherConnections();
             updateIndicator();
             break;
 
-        case 'endpoints':
+        case 'plugin:subscribers': {
             subscribers = packet.data[0];
-            sendToPage(packet);
+            syncPublisherConnections();
             break;
+        }
 
-        case 'getRemoteUI':
-        case 'callback':
-        case 'data':
-            sendToPage(packet);
+        case 'rempl': {
+            const [channelId, payload] = packet.data;
+
+            switch (payload.type) {
+                case 'getRemoteUI':
+                case 'callback':
+                case 'data': {
+                    sendToPage(channelId, payload);
+                    break;
+                }
+            }
+
             break;
+        }
 
         default:
             // @ts-expect-error type
@@ -135,47 +170,110 @@ addEventListener('message', (e: MessageEvent<EventTransportMessage>) => {
     switch (data.to) {
         case connectTo:
             if (data.payload.initiator === connectToName) {
-                onConnect(data.from, data.payload);
+                onPageConnect(data.from, data.payload);
             }
             break;
 
         case inputChannelId:
-            onData(data.payload);
+            onPageDataMessage(data.from, data.payload);
             break;
     }
 });
 
-function onConnect(from: EventTransportChannelId, payload: EventTransportHandshakePayload) {
-    outputChannelId = from;
+let syncPluginPublishersScheduled: Promise<void> | null = null;
+function schedulePluginPublishersSync() {
+    if (!syncPluginPublishersScheduled) {
+        const task = (syncPluginPublishersScheduled = Promise.resolve().then(() => {
+            if (task === syncPluginPublishersScheduled) {
+                sendToPlugin({
+                    type: 'page:publishers',
+                    data: [[...publishers.values()]]
+                });
+            }
+        }));
+    }
+}
+function cancelPluginPublishersSync() {
+    syncPluginPublishersScheduled = null;
+}
 
+function updatePublisherHub(channelId: EventTransportChannelId, publisherNames: string[]) {
+    const newPublishers = [];
+    let publishersChanged = false;
+    let publisherHub = publisherHubs.get(channelId);
+
+    if (!publisherHub) {
+        publisherHub = {
+            id: publisherHubs.size,
+            connected: false,
+            channelId,
+            publishers: []
+        };
+
+        publisherHubs.set(channelId, publisherHub);
+    }
+
+    for (const name of publisherNames) {
+        const publisherId = `${publisherHub.id}:${name}`;
+        let publisher = publishers.get(publisherId);
+
+        if (!publisher) {
+            publishersChanged = true;
+            publishers.set(
+                publisherId,
+                (publisher = {
+                    id: publisherId,
+                    channelId,
+                    name
+                })
+            );
+        }
+
+        newPublishers.push(publisher);
+    }
+
+    for (const publisher of publisherHub.publishers) {
+        if (!newPublishers.includes(publisher)) {
+            publishersChanged = true;
+            publishers.delete(publisher.id);
+        }
+    }
+
+    if (publishersChanged) {
+        publisherHub.publishers = newPublishers;
+        syncPublisherConnections();
+        schedulePluginPublishersSync();
+    }
+}
+
+function onPageConnect(from: EventTransportChannelId, payload: EventTransportHandshakePayload) {
     if (!payload.inited) {
         handshake(true);
     }
 
     remplConnected = true;
-    publishers = payload.endpoints;
     updateIndicator();
 
+    updatePublisherHub(from, payload.endpoints);
+
     if (pluginConnected) {
+        cancelPluginPublishersSync();
+        syncPublisherConnections();
         sendToPlugin({
             type: 'page:connect',
-            data: [sessionId, payload.endpoints || publishers]
-        });
-        sendToPage({
-            type: 'connect',
-            endpoints: subscribers
+            data: [sessionId, [...publishers.values()]]
         });
     }
 }
 
-function onData(payload: EventTransportMessagePayload) {
+function onPageDataMessage(from: EventTransportChannelId, payload: EventTransportMessagePayload) {
     if (DEBUG) {
         console.log('[rempl][content script] page -> plugin', payload); // eslint-disable-line no-console
     }
 
     switch (payload.type) {
         case 'endpoints':
-            publishers = payload.data[0];
+            updatePublisherHub(from, payload.data[0]);
 
             if (!pluginConnected) {
                 return;
@@ -184,7 +282,10 @@ function onData(payload: EventTransportMessagePayload) {
             break;
     }
 
-    sendToPlugin(payload);
+    sendToPlugin({
+        type: 'rempl',
+        data: [from, payload]
+    });
 }
 
 handshake(false);